moving states to State.py
