+++ backend/my_agent/policy_service.py
@@ -61,6 +61,7 @@ from pydantic import BaseModel, Field
 # Import required components
 from backend.my_agent.database_manager import DatabaseManager
 from backend.my_agent.patient_collector import AgentState, PatientCollectorConfig
+from .State import AgentState, create_agent_state
 
 _ = load_dotenv(find_dotenv())  # read local .env file
