+++ config/config.yaml
@@ -2,12 +2,12 @@
 
 models:
   agent_models:
-    - id: meta-llama/llama-4-maverick-17b-128e-instruct
-      provider: groq  # Slow
-    - id: meta-llama/llama-4-scout-17b-16e-instruct
-      provider: groq  # Slow
     - id: mistral-saba-24b
       provider: groq
+    - id: meta-llama/llama-4-maverick-17b-128e-instruct # not stable for structured output
+      provider: groq  
+    - id: meta-llama/llama-4-scout-17b-16e-instruct # not stable for structured output
+      provider: groq  
     # - id: llama3-8b-8192
     #   provider: groq
   tool_models:
