+++ tests/regression/test_end_to_end.py
@@ -19,8 +19,8 @@ def test_end_to_end_workflow():
     try:
         # Import all necessary components
         # from backend.my_agent.llm_manager import LLMManager
-        from backend.my_agent.patient_collector import (  # AgentState,
-            create_agent_state,
+        from backend.my_agent.State import create_agent_state
+        from backend.my_agent.patient_collector import (
             patient_collector_node,
         )
         from backend.my_agent.policy_service import (
