+++ tests/regression/test_trial_search.py
@@ -18,7 +18,7 @@ from dotenv import find_dotenv, load_dotenv
 # sys.path.append(os.path.dirname(__file__))
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "..")))
 
-from backend.my_agent.patient_collector import create_agent_state
+from backend.my_agent.State import create_agent_state
 from backend.my_agent.trial_service import trial_search_node
 
 # Load environment variables
